# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meeting workspace",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/meetings": {
      "get": {
        "summary": "List Meetings",
        "operationId": "list_meetings_api_v1_meetings_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "participant",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Participant"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "minimum",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 21600000,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum"
            }
          },
          {
            "name": "maximum",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 21600000,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maximum"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "seeded",
                    "pasted",
                    "uploaded",
                    "manual"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "recent",
                "oldest",
                "title"
              ],
              "type": "string",
              "default": "recent",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 1000
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "x-demo-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Meeting_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create",
        "operationId": "create_api_v1_meetings_post",
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 100,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-demo-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meeting"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meetings/{meeting_id}/segments/{segment_id}/impact": {
      "get": {
        "summary": "Impact",
        "operationId": "impact_api_v1_meetings__meeting_id__segments__segment_id__impact_get",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Segment Id"
            }
          },
          {
            "name": "x-demo-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentImpact"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/meetings/{meeting_id}/segments/{segment_id}": {
      "patch": {
        "summary": "Edit",
        "operationId": "edit_api_v1_meetings__meeting_id__segments__segment_id__patch",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Segment Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "name": "x-demo-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Segment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Remove",
        "operationId": "remove_api_v1_meetings__meeting_id__segments__segment_id__delete",
        "parameters": [
          {
            "name": "meeting_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Meeting Id"
            }
          },
          {
            "name": "segment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Segment Id"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "name": "x-demo-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Demo-Session"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InlineHighlight": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "start_offset": {
            "type": "integer",
            "title": "Start Offset"
          },
          "end_offset": {
            "type": "integer",
            "title": "End Offset"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "start_offset",
          "end_offset",
          "color"
        ],
        "title": "InlineHighlight"
      },
...
      "Meeting": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "title": "Occurred At"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "source": {
            "type": "string",
            "enum": [
              "seeded",
              "pasted",
              "uploaded",
              "manual"
            ],
            "title": "Source"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "media_mode": {
            "type": "string",
            "enum": [
              "simulated",
              "sample"
            ],
            "title": "Media Mode"
          },
          "sample_media_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Media Key"
          },
          "estimated_timing": {
            "type": "boolean",
            "title": "Estimated Timing"
          },
          "transcript_revision": {
            "type": "integer",
            "title": "Transcript Revision"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/Participant"
            },
            "type": "array",
            "title": "Participants",
            "default": []
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "occurred_at",
          "duration_ms",
          "source",
          "description",
          "media_mode",
          "sample_media_key",
          "estimated_timing",
          "transcript_revision",
          "version"
        ],
        "title": "Meeting"
      },
      "MeetingCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "duration_ms": {
            "type": "integer",
            "maximum": 21600000.0,
            "minimum": 1000.0,
            "title": "Duration Ms"
          },
          "source": {
            "type": "string",
            "enum": [
              "pasted",
              "manual",
              "uploaded"
            ],
            "title": "Source",
            "default": "manual"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/PersonInput"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Participants"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentInput"
            },
            "type": "array",
            "maxItems": 3000,
            "minItems": 1,
            "title": "Segments"
          },
          "tag_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tag Ids"
          },
          "estimated_timing": {
            "type": "boolean",
            "title": "Estimated Timing",
            "default": false
          },
          "acknowledge_estimated": {
            "type": "boolean",
            "title": "Acknowledge Estimated",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "occurred_at",
          "duration_ms",
          "segments"
        ],
        "title": "MeetingCreate"
      },
...
      "Page_Meeting_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Meeting"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          },
          "applied_filters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Applied Filters",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "Page[Meeting]"
      },
...
      "Participant": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "normalized_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized Email"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "normalized_email",
          "version"
        ],
        "title": "Participant"
      },
...
      "PersonInput": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 254,
                "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "PersonInput"
      },
...
      "Segment": {
        "properties": {
          "public_id": {
            "type": "string",
            "format": "uuid",
            "title": "Public Id"
          },
          "speaker_id": {
            "type": "string",
            "format": "uuid",
            "title": "Speaker Id"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "start_ms": {
            "type": "integer",
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "speaker_name": {
            "type": "string",
            "title": "Speaker Name"
          },
          "stable_color_key": {
            "type": "integer",
            "title": "Stable Color Key"
          },
          "highlights": {
            "items": {
              "$ref": "#/components/schemas/InlineHighlight"
            },
            "type": "array",
            "title": "Highlights"
          }
        },
        "type": "object",
        "required": [
          "public_id",
          "speaker_id",
          "ordinal",
          "start_ms",
          "end_ms",
          "text",
          "version",
          "speaker_name",
          "stable_color_key"
        ],
        "title": "Segment"
      },
      "SegmentImpact": {
        "properties": {
          "highlights": {
            "type": "integer",
            "title": "Highlights"
          },
          "comments": {
            "type": "integer",
            "title": "Comments"
          },
          "citations": {
            "type": "integer",
            "title": "Citations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "highlights",
          "comments",
          "citations"
        ],
        "title": "SegmentImpact"
      },
      "SegmentInput": {
        "properties": {
          "speaker": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Speaker"
          },
          "start_ms": {
            "type": "integer",
            "maximum": 21600000.0,
            "minimum": 0.0,
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "maximum": 21600000.0,
            "exclusiveMinimum": 0.0,
            "title": "End Ms"
          },
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "speaker",
          "start_ms",
          "end_ms",
          "text"
        ],
        "title": "SegmentInput"
      },
      "SegmentUpdate": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 10000,
            "minLength": 1,
            "title": "Text"
          },
          "start_ms": {
            "type": "integer",
            "maximum": 21600000.0,
            "minimum": 0.0,
            "title": "Start Ms"
          },
          "end_ms": {
            "type": "integer",
            "maximum": 21600000.0,
            "exclusiveMinimum": 0.0,
            "title": "End Ms"
          },
          "remove_affected_highlights": {
            "type": "boolean",
            "title": "Remove Affected Highlights",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "start_ms",
          "end_ms"
        ],
        "title": "SegmentUpdate"
      },
...
      "Tag": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "name",
          "color",
          "id",
          "version"
        ],
        "title": "Tag"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}